Replace acid-state with hand-rolled PostgreSQL vendoring solution
Ultimately, we need to go further --- the number one problem with the
current deployed (i.e. before this change) Hackage server is the high
resident memory usage due to `acid-state` rolling up the log in memory.

This commit purposely keeps the same architecture of on-disk log and
in-memory roll-up (periodically committed), ad this has all the same
problems.

So why do this at all? Well the answer is that there is still value
getting everything in PostgreSQL at once, and in having a cleaner
separation between in-memory and on-disk data formats.

Once this is done, we can incrementally convert parts of the state to be
stored in a conventional rather than event-sourcing-based pattern, until
the entire app is a normal SQL app with normal access patterns.

+++ src/Distribution/Client.hs
@@ -199,7 +199,7 @@ readNewIndex cacheDir = do
           pkgid
           (Just time)
           (if null username then Nothing else Just (UserName username))
-          (if userid == 0   then Nothing else Just (UserId userid))
+          (if userid == 0   then Nothing else Just (UserId (fromIntegral userid)))
       where
         time     = epochTimeToUTC (Tar.entryTime entry)
         username = Tar.ownerName (Tar.entryOwnership entry)
