+++ src/Distribution/Client/Mirror/Repo/Util.hs
@@ -44,7 +44,7 @@ readIndex repDescr indexFile = do
           pkgid
           (Just time)
           (if null username then Nothing else Just (UserName username))
-          (if userid == 0   then Nothing else Just (UserId userid))
+          (if userid == 0   then Nothing else Just (UserId (fromIntegral userid)))
       where
         time     = epochTimeToUTC (Tar.entryTime entry)
         username = Tar.ownerName (Tar.entryOwnership entry)
